requirements.txt
streamlit
transformers
torch
torchvision
pillow
numpy
opencv-python-headless
scipy
